Factor out CfJsonDocumentContext
